backend1/requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.5
django-phonenumber-field==8.1.0
gunicorn==23.0.0
packaging==25.0
phonenumbers==9.0.15
psycopg2-binary==2.9.11
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
whitenoise==6.11.0
